+++ crates/p2p/src/commands.rs
@@ -3,7 +3,7 @@
 use libp2p::{Multiaddr, identity::PublicKey};
 use tokio::sync::oneshot;
 
-use crate::swarm::message_dht::RecordData;
+use crate::swarm::dht_record::RecordData;
 
 /// Commands that users can send to the P2P node.
 #[derive(Debug)]
